[FIX] qweb: translate: allow looking up translation in parent template
Translations lookup normally uses the namespace of the current
QWeb template, after merging all inherited views.
But when a QWeb template is "cloned" by a child view using
inheritance with `primary` mode, the translations are more
likely to exist for the original (parent) template, and would not
be found when using only the "child" namespace.

This patch adds support for looking up each translation
also in the parent namespace in this case, if none was found
for the child template in the first place.