[IMP] Web: action_manager.js: attach/detach actions
When several actions are stacked by the ActionManager, only the last
(current) one is in the DOM, the previous ones are detached and stored
to allow restoring them later (by clicking on the breadcrumbs).

The cleans the DOM concerning the ActionManager.

Moreover, when switching back to an action by clicking on the
breadcrumbs, the further actions are destoyed only when the clicked one
is ready to be displayed. The switch is thus now done atomically.