[FIX] web: misc: monkey-patch nvd3's tooltips handler
In order to set a tooltip's highest position in the graph zone and not
over the control panel.

[FIX] webclient layout: 100% height + distribute it to o-view-manager-content

Done with a css hack, the flex item `o-content` has now a height 100%
(inherited from its parent `o-web-client` and, in order to distribute it
to its child `o-view-manager-content`, set its display to relative. This
way, its child can set its display to absolute and thus extend itself to
the limits of its parent. Obviously, it is done because the traditional
height 100% inheritance is broken with growing flex item.

* graph view are using this mechanism
* pivot view are using this mechanism
* kanban view are using this mechanism in a wrong way (min-height 100%,
  so that the empty colum take 100% height if its sibling columns are not
  deep enough and the size of its sibling column it they are deep
  enough).

[REF] web: layout: remove grid-gutter-width

and set manually the padding as in most cases we don't want it/we want
to fine tune it.

[FIX] web: navbar: menu brand does not open the app switcher