[IMP] sale : discount related improvement on the quotation reports
The quotation report has been improved to show the discount on lines
when available, with the currency symbol, etc.

Also a new 'price_discounted' field has been added to the
'sale.order.line':

The price_discounted functional field gives the unit price of the
product, after the discount has been applied. It is there to replace
ad-hoc computation, and indicate the correct rounding precision,
for when it needs to be displayed in the web client or in reports.