[FIX] point_of_sale: ignore unrecognized barcodes in popups
When a popup is displayed in the point of sale, it sets-up its
own barcode callbacks to prevent the screen's default barcode
action. If a barcode doesn't handle such a barcode, and has not
been setup to handle errors either, it will be ignored, and
a warning will be emitted to the console.