Deal with errors differently
The _required_information_filled_in function should mainly be used to
check for fields that we need in the receipts.

For other fields that we need directly in OS, we should just be able to
raise exceptions when we need them. That way we don't have some weird
disconnect between "checking if it's filled in" and using it.

We need some values pretty deep in the call stack, so returning false
all the way to the top doesn't work. So we need to throw an
exception. If you don't catch an exception in the POS it'll show a
backtrace, which we don't want. So we need to catch it. Issue is that js
doesn't support conditional catches (except in Firefox) so just
differentiate based on a message string prefix I guess.