Also save base amount per tax category to pos_order
There is also a rename eurocents -> eurocents in here. Furthermore we
made sure that the base amount per tax category is only calculated
once. Before this we would do multiple calls of
get_base_price_in_eurocents_per_tax_letter() etc. Doing only one call
and saving it on order.blackbox_base_price_in_euro_per_tax_letter
guarantees that these values are always the same on the FDM, on the
receipt and in the backend.