Temporarily just wait for ACKs from the blackbox device when probing
I'd like this to be temporary because it would be more robust. Before
this we would parse the entire response (including checksum) so the odds
of a non-blackbox device being picked as one was non-existant. Right now
we only listen for an ACK (0x13) after our probe request, which isn't as
robust (maybe there's a eg. scale that happens to broadcast a weight of
0x13 every second).

Anyway, the reason we did this is that when the blackbox receives a
duplicate request I think it just ACKs without sending an actual
response. Which kind of ruins the whole probing process. I'll ask the
government about this next week.