Fix the way we check when orders need to be send to the backend and FDM
We have to ensure that when we send an order to the backend we also send
it to the FDM. This is a bit tricky because we need information from the
FDM to put on the receipt (which is printed after the order is sent to
the backend). On top of that the rpc stuff uses an asynchronous call,
which makes a bit trickier.

The solution is to patch push_order() to do our own blackbox related
validation. Because of this the order is first validated by
posmodel.validate_order(). Then we send it to the FDM, then we put the
response fields on the order, then we print the receipt and lastly we
send it to the backend.

This still won't work when we invoice a POS order because then no
receipt will be printed. I think the solution there should just be to
print a valid VAT receipt anyway. Otherwise I have to put all the extra
FDM related fields on the invoice as well which is kind of a pain to
do. So in that case you would give the customer an invoice + VAT
receipt.