Add amount_total to pro_forma orders
So for pro_forma we will just use the amount calculated by the
client. Normally, we recompute everything (total amount, taxes, ...) in
the backend (and then hope that the result is the same as in the
client). But we'd have to copy over all the compute functions from
pos_order that do these computations, which I don't want to do. So
instead we just send the total amount, which will be used to set the
amount_total float in pos_order_pro_forma.

Also add the total revenue from pos_order_pro_forma in a pos_session to
the report.