Never allow changes that decrease the total price of an order
After talking to Yves, we have to make sure that people cannot affect an
order in such a way that it decreases the total price without there
being any indication of what actually happened. So this is *always*
necessary, not just for orders that have been pro forma'd. A practical
example of is eg. when a user misclicks the wrong product, the only way
to correct this is by adding an orderline with the same product and a
negative quantity. An added complication is that we have to make sure
that we do not return more than what was ordered in the first place (no
negative quantities or negative monetary amounts can be sent to the
FDM).

So anyway, to deal with this we make it so that users cannot decrease
the quantity of an orderline. When trying to do this the POS will show a
dialog asking by how much (quantity) the user wants to decrease the
orderline. We then test if that decrease is valid (< product quantity
left). After that we create an extra orderline which will contain the
negative quantity.

Users are still able to change the quantity of an orderline as long as
the change is positive.