Add X and Z user reports
In order to do this properly we also printed the proper time on the
ticket (by default it prints american time (AM/PM) and no seconds, which
is weird). And we also returned this time to the backend, this way we
have the 'receipt time' that is sent to the FDM, the backend AND is
printed on the receipt.