Refactor push_order
Get rid of the pro_forma flag. Instead provide a variable
'after_blackbox' which should be a function that will be executed after
we have received the response from the blackbox. This is used for
displaying the "bill" in pos_restaurant and it's also used to mark
orderlines as pro_forma_finalized after pushing a pro forma order.