Round instead of flooring
So here's my justification for it: We multiply by 100. So really we
shouldn't have any fractional part anymore. Flooring seems the more
'correct' thing to do but we can't in this case because sometimes we
have a negative quantity and thus negative price eg. -4.999. That should
clearly be -5, but if you floor you end up with -4 which ends up causing
negative base tax amounts that are sent to the FDM. Rounding takes care
of that issue and I don't think it's going to cause further issues.