[FIX] point_of_sale: don't hardcode an orderline quantity and qty string
Setting a default qty of 1 on an orderline is correct, but it should be
set with the set_quantity function so that it gets properly formatted.

Otherwise you end up with a quantity of 1 when you first add a product,
and when you increase the quantity to 2 it'll turn into 2.000.

opw-657591