[IMP] point_of_sale: automatically set customer fiscal position
This makes it so that when you set a customer in the POS frontend it
will automatically apply the fiscal position associated to the customer
on the order.

If the customer has no set fiscal position or if the fiscal position is
not available in the POS nothing will happen.