[IMP] point_of_sale: allow limit specification during model loading
This allows for significant performance improvements in pos_blackbox_be,
because it needs to load the most recent POS order belonging to the
current POS configuration. Without the limit all POS orders would have
to be loaded.

Although there exist different ways to avoid these performance
issues (RPC or controller), it makes little sense to go through all that
trouble if web.data.Query has the limit functionality and the POS just
has to use it.