[FIX] point_of_sale: maintain stable viewport on iOS
The POS frontend has an unchangeable viewport which should be 1024 px
wide. On iOS 9.2.1 however, this viewport sometimes changes to a much
smaller value (154 px) after fading out the loading screen. This makes
the POS pretty much unuseable, especially considering that this viewport
is not user-scalable, which means you can't zoom out.

This is due to a recent change to WebKit2 for iOS [1] which makes the
viewport resize when setting the overflow property on
document.body. This particular change has been recently reverted [2],
but this does not seem to have made it to the latest stable WebKit build
for iOS as of the time of this commit.

Because of this we will remove the call to set_content_fullscreen
entirely. It only does two things:

1. set overflow-y to hidden on the document.body
   This is reduntant because we already have 'overflow: hidden' on the
  .pos div, which contains the entire application
2. adds .oe_content_full_screen to the .oe_webclient div
   This is also not particularly useful because the stylesheet rules for
   this class are contained in addons/web/static/src/css/base.sass which
   isn't loaded in the POS

[1] https://trac.webkit.org/changeset/186786
[2] https://trac.webkit.org/changeset/194667