[FIX] mail queue: wrong condition for test mode
Rev. 56cec604e6313c49582bbd9801686328696b2675 modifies
the auto-commit behavior of the mail queue scheduler,
to allow using it within tests.

The condition was wrong and always True as soon as
the registry was loaded, permanently disabling
auto-commit, both during tests and outside tests.

This broke the duplicate emails protection.