[FIX] payment_stripe: support payments through website_quote
Contrary to other payment acquirers all transactions through stripe are
s2s transactions. Because of this payment_stripe has to create it's own
payment.transactions which is unusual.

Because of this payment_stripe needs a sale order id which will be
provided by website_sale in request.session. website_quote doesn't do
this so this extracts it from the return_url.

TODO EXPLAIN invoice_num, also test how this ever worked anyway because
invoice_num is always '/' in website_sale and website_quote