[FIX] point_of_sale: avoid scrollbar on main window
Related to depending on non-standard behavior when setting height on a
table. For a detailed explanation see
f0eeed93dc5e8ee8eeb92864c75293de6c082fff.

In this case we can simply ignore overflow on the main POS window since
the height difference introduced is only a couple of pixels.