[FIX] calendar: when calling the model's search, exclude the virtual ids
Calendar's overriden search plays with the domain. It also calls the
original search method and work with the result, to handle the vitual_id
things. The original search will calls again the calendar's search to
find the database ids, (see parse method). In this special case, we must
be sure to short circuit the virtual_id handling, else the generated
query will look for virtual_ids and we'll have a traceback.