[REF] stock: basics of new model and stock.move reservation
We use this commit to reimplement `action_assign` (wiped by the new
quant implentation). A move is reserved if its move lines exist, so we
introduce a one2many between stock.move and stock.pack.operation.
We remove the `pack_operation_pack_ids` and `pack_operation_product_ids`
as they will be represented by the same model, the `pack_operation_ids`
with the necessary tweaks in the views (once again, in a further
commit).