[FIX] P3: round() compatibility shim
In Python 3 round:

* rounds half to even (banker's rounding) rather than up
* returns an int if not given a precision (or given a precision of
  None)
* discards the negative sign when rounding to 0

Add a compatibility shim which implements P2 behaviour and at least
allows passing float_utils's tests, implementing float_round on top of
Decimal was attempted but did not work (only uses exponent when
quantizing to rounding to a precision of e.g. 0.05 does not work as
the tests expect).