[IMP] l10n_co: update and simplify
This simplifies the chart of accounts and taxes to only the common
ones used. It also updates some changes that took place (eg. the VAT
increase from 16% to 19%). Lastly it implements some additional fields
on res.partner: a document type and a verification code that is
calculated based on the VAT.