[FIX] web_studio: not able to set pad widget
Before this fix, it was possible to set the widget 'pad'
on char fields but it only worked if the current model
inherited from 'pad.common'.
In the other cases, it crashed on closing Studio.

For now, there is two models where you can set this widget:
project.task and note.note

As this is specific for two models only (project.task and
note.note), we decided to remove the possibility to set
this widget in Studio.

opw-745997