[IMP] mrp: add a line in the produce wizard and fix
Make it possible to add an extra line in the produce wizard.  This was
done for when we consume more than foreseen to be able to still add e.g.
the serial number.

The lots shown in the wizard, should only be those without final lot
associated.  When done, it should just adapt the quantities, not add
them.