[ADD] web: added method (controller) for setting db before calling module controller in web
Issue: for calling json controller for module if there is no db-filter or db available in session, server not able to find route for that module
and returns 404 error for that controller.

Solution: Added controller in web (/web/db/ensure) which take db name as parameter and set databsae to session.db and after
ensure db controller for that database will work.