[FIX] point_of_sale: update partner search when deleting input
The keypress event is only fired for keys that produce a
character. This does not include backspace. This meant that the
partner selection screen didn't update when backspacing part of a
search query.

To solve this use the input event instead of keypress. This event
fires when the value of the <input> element is changed.