[FIX] account: Always track payments'invoices properly
For this, convert invoice_ids field of account.payment into a computed field.

By consistency, also convert payment_ids of accoun.invoice into a computed field.

To allow this change, I had to change the way the posting of invoice payments was made, as we used to just set the invoice in advance and then post everything, and it does obviously not work anymore. I added an optional parameter to the post() method, allowing to give the invoices for which we are posting the payments.

[WIP: still have to correct the tests, and possibly one or two little bugs]