[IMP] website: Display a specific template when an access error raised
Currently website's template will be displayed to users when access error raised. It's very confusing especially for visitors while getting this message.

Let's take an example of a forum: If there is a rule for model 'forum.forum' with public/portal user regarding website published, archived a forum and now if public user try to go any url(path) regarding forum which contains forum id then they will get '403: Forbidden' message because access error will be raised as per created rule and as per error code website's template will be rendered. This is very confusing for visitors to know what's going wrong.

So, added one new parameter 'error_templates' to route to handle exception with a specific template for a specific error code. This 'error_templates' is a disctionary that contains an error code as a key and a template as a value. At the time of exception occurence, first route will find a 'error_templates' and if it's there then it will render a specific template for a specific error code otherwise it'll render a default template (i.e. website's templates)