[IMP] http: allow fine-tuning of session FS storage
With this commit, it is now possible to fine-tune which kind of sessions
should be stored into the file-system by simply invoking the
`--sessionless=` command-line option and passing it a comma-separated
list of context keys for which session storing should be ignored.

e.g.: `--sessionless=uid,lang,tz` means that any session with a context
whose keys are a subset of {uid, lang, tz} won't be stored in the FS.

The purpose of this feature is to reduce performance issues related to
these session files cluttering the FS, as a lot of these files
are created but never used.