[ADD] account_invoice_import: add generic import e-invoicing + Factur-x standard
Factur-x is the e-invoicing standard in vigor in FR & DE. However, this will be used as
default e-invoicing standard by Odoo. The idea is simple: when printing your report
(not in draft ofc), a Factur-x xml file is generated and embedded inside the PDF (hidden to the
user then). When importing the invoice report, the invoice values are pre-filled according to
the embedded xml.

The import could be done either by clicking on 'Import Bills' in the dashboard
(under purchase journals) or either by configuring a mail alias in the purchase journals and then,
update the invoice according to the received attachment by mail.

task: 1859277