[FIX] web, web_editor: restore report edition feature
Since 11.0, reports could not be edited with studio anymore. This is due
to the fact the editor files were refactored but the report feature is
not correctly organized since it was merged with the 'web' app. Indeed,
'web' app is still calling files and assets of the 'web_editor' app which
is not a dependency (but it is working as 'web_editor' is auto installed
with 'web').

opw-1826599
opw-1834971
Closes 24034 !!!!!!!!!!!!!