[IMP]stock: Add record_rule in po,
In user create the record rule for product then generate the po for product that time show the record rule name in po source_document field
Receive products in 2 steps,
The product has a reordering rule,
On the po generated, the source document should be the name of the reordering rule.

Task: 1838639.