[IMP]doc: update documentation on customizing the web client
with this commit, we rework a lot of the web client doc.
in this doc we have create a simple module, which is holding basic web component configuration and letting us test the web framework.
explained little bit information of odoo JavaScript Module, Classes, Widget, Translation and how to communicate with the odoo server