[CLEAN-FIXME] gamification: Removed fixme comment
A model decorator is necessary because a cron function also called from a 'website_forum' module through function tag in demo data file. If we remove decorator then it'll take it as a multi type function and called 'call_kw_multi' function which is finding the zero indexed args and in that case we'll get 'list index out of range' as we've blanked list.

Related to Task: 58786