[FIX] web: data_manager: handle load_action rpc errors
Before this commit, load_action was never resolved in case of
internet connection loss.
It's because the call is triggered by web.rpc direcly who uses
ajax.rpc.

By executing this call by web.session, we take the benefits from
the error checks if the response fails .

We didn't change the other calls to minimize the diff and avoid
bugs due to this change.
So, it's why only the first call ("/web/action") is checked.

Steps to reproduce the issue:
- Go to the app switcher for the first time
- Force disconnected from the network
-> No notification appears and the user aren't notify

Note: It's now important to notify the user about connection loss
because a recent improvement in mobile app replaces the 'retry screen'
and let the webclient handle connection issues by itself.
The single odoo mobile app is supposed to work with all databases
from 10.0.