[FIX] web: data_manager: handle load_action rpc errors
Before this commit, load_action was never resolved in case of
internet connection loss.
This is due to the use of web.rpc, which directly uses ajax.rpc
without handling errors.

By executing this call by web.session, we take the benefits from
the error checks if the response fails .

We didn't change the other calls to minimize the diff and avoid
bugs from this change.
So, that's why only the first call ("/web/action") is checked.

Steps to reproduce the issue:
- Go to the app switcher for the first time
- Force disconnected from the network
-> No notification is shown to the user

Note: It's now important to notify the user about connection loss
because a recent improvement in mobile app replaces the 'retry screen'
and lets the webclient handle connection issues by itself.
The odoo mobile app is supposed to work with all databases from 10.0.