[IMP] Mass Mailing - Implement Double Opt In - Send invitation mail
In the email marketing app, add a "double opt-in" option
When enabled : add A double Opt-in option on the mailing lists. True by default. Possibility to disable it individually.
When disabled, removed the parameter on the mailing lists (==> false)
When someone subscribes to a mailing list (for example, from the newletter website snippet), or when he is added to it manually from the backend, an email should be sent to ask if the recipient agrees. Once he agrees, he will be part of the mailing list.
When the option is enabled, on the mailing list contact form: In the mailing list many2many.
Add a Status column : Waiting for confirmation, Confirmed, Opt-out  (Remove the Opt-Out column).
The status change should be logged in the chatter. Who and how (email or manually)
If double opt-in is disabled on the mailing list : Confirmed, Opt-Out
If double opt-in is disabled on the mailing list : Waiting for confirmation become Confirmed. Add a Warning message to inform the user about the changes.
On the mailing list recipients tree view + On the recipient form view: add an action button " Re-send invitation" in case the recipient was opted-out by error. The tooltip = Re-send opt-in invitation
If a recipient declines the invitation, status becomes opt-out + message logged in the chatter (recipient has declined the invitation)
Emails shouldn't be sent to recipients that are "Waiting for confirmation" or Opt-Out.