[FIX]purchase: Issue with product_code and supplier info. Before this commit -
If a product is configured with 2 seller ids with different prices as -

Product - iPhoneX, name - iPhoneX_XX, code - 01, minimum purchase - 1, unit price - 500
Product - iPhoneX, name - iPhoneX_YY, code - 02, minimum purchase - 10, unit price - 300

When we select iPhoneX_XX then desc and price are correct. But when we change the purchase qty -

Price - 300, desc - iPhoneX_XX
which was wrong(desc).

After this commit -
Onchange of purchase qty, best seller will be displayed and according to that desc will come.
This merge is related to bug 1817414.