[FIX] web: JournalDashboardGraph widget crash
same goal as 995610c065bf0242cb5023cfad2940234395739e , just one little step further

commit message copied from 995610c065bf0242cb5023cfad2940234395739e :
"
The JournalDashboardGraph requires nv, which is lazyloaded. Before
this rev., a crash occurred when this widget was instantiated and
destroyed before the loading of the lib was complete (because a
function of the lib was called in destroy()).

For instance, press F5 (to ensure that nv isn't already loaded),
activate some throttling in the network tab, go to Accounting and
as soon as the dashboard shows up, click on another menu."

ticket ref:  # 1873749