SVG: Include mimetype detection for SVG
When displaying an image through the controller triggered by /web/image/* URLs, a
call for guessing the mimetype is done, defaulting on "image/png" is no other guess
is done.

This works for common raster image formats (PNG, JPG, GIF), but not for SVG. Adding
a simple mimetype entry that detects the xml beginning plus a discriminant that
checks that it has the svg opening and ending tags.

There's no possible risk in this operation, as we only makes the checks through
plain text.